rename indexed-formula.js to store.js . A good idea anyway but does not fix the jsdoc gen problem
