+++ package.json
@@ -66,7 +66,7 @@
   "scripts": {
     "build": "babel src -d lib",
     "build:browser": "webpack --progress",
-    "doc": "jsdoc -d doc README.md src/*.js",
+    "doc": "jsdoc -d doc README.md src/*.js && mv doc/global.html#IndexedFormula doc/IndexedFormula.html",
     "prepare": "npm run build && npm run build:browser",
     "test": "npm run test:unit && npm run test:serialize",
     "test:clean": "rimraf tests/serialize/,*",
